\name{model}
\alias{model}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Model selection based on credible intervals}
\description{
Model selection based on credible intervals}
\usage{
model(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{ an output object of the Brq function.}
  }
