\name{print.Brq}
\alias{print.Brq}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print the Bayesian quantile regression methods
}
\description{
  \code{print.Brq} is an S3 method that prints the content of an S3 object of class \code{Brq} to the console. 
}
\usage{
%print.Brq(x, ...)
\method{print}{Brq}(x,\ldots)
}
\arguments{
  \item{x}{
 \code{object of class Brq} 
}
  \item{\dots}{
\code{not used.} 
}
}
\value{
Print the Bayesian quantile regression methods
}
