\name{salary}
\alias{salary}
\docType{data}
\title{Salaries of 459 US statistics professor data}
\description{A university union collected data on the salaries of 459 US statistics professors and the 
number of years for which they had been professors during the period from 1980 to 1990 (Bailar, 1991).}
\usage{data("salary")}
\format{
  A data frame with 459 observations on the following 2 variables.
  \describe{
    \item{\code{Salaries}}{ Salaries of 459 US statistics professor (1000s of dollars).}
    \item{\code{Years}}{ Number of years as professor}
}
}
\source{
[1] Bailar, B. (1991) Salary survey of U.S. colleges and universities offering degrees in statistics. Amstat News, 182, 
3-10. 
}

\references{
[1] Bailar, B. (1991) Salary survey of U.S. colleges and universities offering degrees in statistics. Amstat News, 182, 
3-10. 
}

\keyword{salary}
