\name{USgirl}
\alias{USgirl}
\docType{data}
\title{ Weight and age for a sample of 4011 US girl data}
\description{This data set comprises weight  and age for a sample of 4011 US girls reported in  Cole (1988) and used in Keming et al. (2003).
}
\usage{data(USgirl)}
\format{
  A data frame with 4011 observations on the following 2 variables.
  \describe{
    \item{\code{Age}}{Age (years)}
    \item{\code{Weight}}{Weight (Kg)}
  }
}

\references{
[1] Cole, T. J. (1988) Fitting smoothed centile curves to reference data (with discussion). J  R. Statist. Soc. A, 151, 
385-418.

[2] Yu, K., Lu, Z., and Stander, J. (2003). Quantile regression: applications and current research areas. Journal of the Royal Statistical Society: Series D (The Statistician), 52(3), 331-350.
}

\keyword{USgirl}
