\name{wheat}
\alias{wheat}
\docType{data}
\title{Wheat Data}
\description{
This is a regression data set consisting of 584 observations on 11 variables is taken from the national program for the development of wheat cultivation in Iraq - Qadisiyah governorate branch (2017). The outcome of interest in this dataset is the percentage increase in wheat yield per dunam (2500 square meters).
}
\usage{data("wheat")}
\format{
  A data frame with 584 observations on the following 11 variables.
  \describe{
    \item{\code{y}}{The percentage increase in wheat yield per dunam (2500 square meters).
}
    \item{\code{U}}{Fertilizer urea (a numeric variable coding the quantity of fertilizer in kilogram).}
    \item{\code{Ds}}{The date of sowing wheat seeds (a numeric variable coding date: 1 the ideal date, 2 early date, 3 late date).}
    \item{\code{Qs}}{ The quantity of sowing wheat seeds (a numeric variable coding the quantity of seeds in kilogram).}
    \item{\code{LT}}{The laser field leveling technique (a numeric variable coding technique: 2 if there are used this technique; 1 otherwise; "LT").}
    \item{\code{NPK}}{The fertilize the field with compound fertilizers (a numeric variable coding the quantity of fertilizer in kilogram).}
    \item{\code{SMT}}{Seed sowing machine technique (a numeric variable coding the seed sowing machine technique: 2 if there are used this technique; 1 otherwise).}
    \item{\code{SC}}{ Planting  mung bean crop (a numeric variable coding type: 2 planting mung bean, 1 otherwise).}
    \item{\code{H}}{Herbicide for weed control in crops (a numeric variable coding the quantity of herbicide in milliliter).}
    \item{\code{K}}{High potassium fertilizer for crops (a numeric variable coding the quantity of fertilizer in kilogram).}
    \item{\code{ME}}{Microelement fertilizer (numeric variable coding the quantity of fertilizer in gram).}
  }
}

\keyword{datasets}
