\name{DIC}
\alias{DIC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Estimate Deviance Information Criterion (DIC)}
\description{
This function computes the Deviance Information Criterion (DIC),  which is a hierarchical modeling generalization of the Akaike Information Criterion. 
It is useful for Bayesian model selection.}
\usage{
DIC(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{ an output object of the Brq function.}
  }
