\name{wheat}
\alias{wheat}
\docType{data}
\title{Wheat Data}
\description{
This is a regression subsample data set consisting of 113 observations on 6 variables is taken from the national program for the development of wheat cultivation in Iraq - Qadisiyah governorate branch (2017). The complete data set was first analyzed by Hilali (2019). The outcome of interest in this dataset is the percentage increase in wheat yield per dunam (2500 square meters).}
\usage{data("wheat")}
\format{
  A data frame with 113 observations on the following 6 variables.
  \describe{
    \item{\code{y}}{a numeric vector}
    \item{\code{U}}{a numeric vector}
    \item{\code{Qs}}{a numeric vector}
    \item{\code{H}}{a numeric vector}
    \item{\code{K}}{a numeric vector}
    \item{\code{ME}}{a numeric vector}
  }
}
\references{Abbas, H. K. and Taher, R. J. (2019). Bayesian adaptive lasso tobit regression. \emph{Journal of Al-Qadisiyah for Computer Science and Mathematics}, 110 (1). <doi:10.29304/jqcm.2019.11.1.471>
}
\keyword{datasets}